refactor: 초기 로드하는 지도 커스텀 훅 분리 #92
